+++ src/server/devices/deviceType.ts
@@ -42,4 +42,5 @@ export enum DeviceType {
   Espresense = 4001,
   TrackableDevice = 4002,
   SamsungTv = 5001,
+  Camera = 6001,
 }